requirements.txt
asgiref==3.9.1
boto3==1.35.84
botocore==1.35.99
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
Django==5.2.6
django-cors-headers==4.7.0
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
jmespath==1.0.1
openai==1.3.7
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-magic-bin==0.4.14
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
reportlab==4.0.4
requests==2.32.5
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
stripe==11.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.32.1
whitenoise==6.9.0
